DEV-3070 Resolve PR comments WIP
