+++ pom.xml
@@ -37,7 +37,6 @@
         <org-reflections.version>0.9.11</org-reflections.version>
         <google-api-services-serviceusage.version>v1beta1-rev20200804-1.30.10</google-api-services-serviceusage.version>
         <events.version>1.0.0</events.version>
-<!--        <java-client.version>1.12.0</java-client.version>-->
         <java-client.version>1.31.1</java-client.version>
         <compar.version>0.0.3</compar.version>
         <org-reflections.version>0.9.11</org-reflections.version>
